+++ NEWS.md
@@ -14,3 +14,4 @@ Version 0.4.0 - xx Oct 2013
   for MAC's, do the same here
 * fix protocol field en/decoding for route events
 * de/encode getroute and getneigh events
+* netlink cache for neighbour's and route's